Adding CPU-GPU async
The idea is that the laguerre calculation should run in paralel with
everything happening in cuda. This commit adds such an infrastructure,
relying mostly on new class to hold the computation's state.
Note however that this change does not increase the calculation speed so
further optimisation is neede. Also, python calculation fails for
unknown reason.